+++ HISTORY.rst
@@ -7,9 +7,9 @@ Release History
 
 - Added support for AMQP Sequence as the body type of the message.
 - Added new class `uamqp.MessageBodyType` to denote the body type of an amqp message, including:
-  - `DataType`: A data section contains opaque binary data.
-  - `SequenceType`: A sequence section contains an arbitrary number of structured data elements.
-  - `ValueType`: An amqp-value section contains a single AMQP value.
+  - `DataType`: The body consists of one or more data sections and each section contains opaque binary data.
+  - `SequenceType`: The body consists of one or more sequence sections and each section contains an arbitrary number of structured data elements.
+  - `ValueType`: The body consists of one amqp-value section and the section contains a single AMQP value.
 - Added new parameter `body_type` to the constructor of `uamqp.Message` which takes `uamqp.MessageBodyType` to denote the body type of an amqp message.